integrate InputAction and ConsoleLogger
